+++ bigmler/__init__.py
@@ -1,2 +1,2 @@
 # -*- coding: utf-8 -*-
-__version__ = '3.14.0'
+__version__ = '3.14.1'